# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.SystemResources:
    properties:
      cpu_usage:
        description: percent (0-100)
        type: number
      disk_total:
        description: bytes
        type: integer
      disk_used:
        description: bytes
        type: integer
      load1:
        type: number
      load5:
        type: number
      load15:
        type: number
      memory_total:
        description: bytes
        type: integer
      memory_used:
        description: bytes
        type: integer
      uptime:
        description: seconds
        type: integer
    type: object
host: localhost:8090
info:
  contact:
    name: PopuGate Support
    url: https://github.com/fussraider/PopuGate
  description: PopuGate is a Telegram MTProto proxy manager with a REST API.
  license:
    name: MIT
    url: https://github.com/fussraider/PopuGate/blob/main/LICENSE
  title: PopuGate API
  version: "1.0"
paths:
...
  /system/os:
    get:
      consumes:
      - application/json
      description: Returns detected OS information including distribution, version,
        and architecture
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Detect operating system
      tags:
      - system
  /system/resources:
    get:
      consumes:
      - application/json
      description: Returns current CPU, memory, disk, and load information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SystemResources'
      security:
      - BearerAuth: []
      summary: Get system resources
      tags:
      - system
  /system/service/install:
    post:
      consumes:
      - application/json
      description: Installs and enables the PopuGate systemd service unit
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Install systemd service
      tags:
      - system
  /system/service/reload:
    post:
      consumes:
      - application/json
      description: Sends a reload signal to the PopuGate systemd service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reload service
      tags:
      - system
  /system/service/restart:
    post:
      consumes:
      - application/json
      description: Restarts the PopuGate systemd service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Restart service
      tags:
      - system
  /system/service/status:
    get:
      consumes:
      - application/json
      description: Returns the current systemd service status (running, stopped, etc.)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Service status
      tags:
      - system
  /system/service/uninstall:
    delete:
      consumes:
      - application/json
      description: Stops and removes the PopuGate systemd service unit
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Uninstall systemd service
      tags:
      - system
...
  /update/apply:
    post:
      consumes:
      - application/json
      description: 'Downloads and installs the latest PopuGate update, then triggers
        a restart. In binary mode: replaces the binary. In Docker mode: pulls a new
        image and recreates the container.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Apply PopuGate update
      tags:
      - update
  /update/check:
    get:
      consumes:
      - application/json
      description: Checks remote source for new PopuGate releases and returns update
        availability status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Check for PopuGate updates
      tags:
      - update
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"